requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
autopep8==2.0.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.2.1
et-xmlfile==1.1.0
Flask==2.2.2
frozenlist==1.4.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.4
numpy==1.26.1
openai==0.28.1
openpyxl==3.0.10
pandas==2.1.2
pandas-stubs==1.5.2.221213
pycodestyle==2.10.0
PyJWT==2.6.0
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
six==1.16.0
tomli==2.0.1
tqdm==4.64.1
twilio==8.10.0
types-pytz==2022.6.0.1
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
Werkzeug==2.2.2
yarl==1.9.2
zipp==3.11.0
